dbx_bundle/pyproject.toml
[project]
name = "dbx_bundle"
version = "0.0.1"
authors = [{ name = "yoichiojima@gmail.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest",
    # Code completion support for Lakeflow Declarative Pipelines, also install databricks-connect
    "databricks-dlt",
    # databricks-connect can be used to run parts of this project locally.
    # Note that for local development, you should use a version that is not newer
    # than the remote cluster or serverless compute you connect to.
    # See also https://docs.databricks.com/dev-tools/databricks-connect.html.
    "databricks-connect>=15.4,<15.5",
    "databricks-sdk>=0.65.0",
    "jupyterlab>=4.4.9",
    "ruff>=0.13.2",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dbx_bundle"]

[project.scripts]
main = "dbx_bundle.main:main"
